Refactor agent message renderer modules
