+++ src/features/agent/components/AgentMessageRenderer/types.ts
@@ -1,4 +1,4 @@
-import type { MCPContent } from '@/lib/mcp';
+import type { MCPContent, MCPToolCallContent } from '@/lib/mcp';
 import type { Message } from '@/models/chat';
 
 export interface AgentMessageRendererProps {
@@ -10,6 +10,13 @@ export interface AgentMessageRendererProps {
   /** Map of tool call ID to result message (for unified rendering) */
   toolResultsMap?: Map<string, Message>;
 }
+
+export interface ToolGroupBlock {
+  type: 'tool_group_block';
+  items: MCPToolCallContent[];
+}
+
+export type RenderItem = MCPContent | ToolGroupBlock;
 
 // Helper type to avoid implicit any in markdown components
 export type MarkdownReflessProps<T extends React.ElementType> =
